apps/gateway/requirements.txt
litellm[proxy]==1.77.5
PyYAML==6.0.2
requests==2.32.3
